add integration test to demp-rollup
