+++ examples/demo-rollup/src/main.rs
@@ -1,84 +1,26 @@
 use std::env;
-use std::str::FromStr;
-use std::sync::Arc;
 
 use anyhow::Context;
 use borsh::{BorshDeserialize, BorshSerialize};
-use const_rollup_config::{ROLLUP_NAMESPACE_RAW, SEQUENCER_DA_ADDRESS};
-use demo_stf::app::{App, DefaultContext, DefaultPrivateKey};
-use demo_stf::genesis_config::create_demo_genesis_config;
-use demo_stf::runtime::{get_rpc_methods, GenesisConfig};
+use demo_stf::app::{App, DefaultContext};
+use demo_stf::runtime::get_rpc_methods;
 use jupiter::da_service::CelestiaService;
 #[cfg(feature = "experimental")]
 use jupiter::da_service::DaServiceConfig;
-use jupiter::types::NamespaceId;
-use jupiter::verifier::address::CelestiaAddress;
 use jupiter::verifier::{ChainValidityCondition, RollupParams};
 use risc0_adapter::host::Risc0Verifier;
-use sov_db::ledger_db::LedgerDB;
+use sov_demo_rollup::register_rpc::{register_ledger, register_sequencer};
+use sov_demo_rollup::{get_genesis_config, initialize_ledger, ROLLUP_NAMESPACE};
 #[cfg(feature = "experimental")]
 use sov_ethereum::get_ethereum_rpc;
-use sov_modules_stf_template::{SequencerOutcome, TxEffect};
-use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::services::da::DaService;
 use sov_rollup_interface::zk::ValidityConditionChecker;
-use sov_sequencer::get_sequencer_rpc;
 use sov_state::storage::Storage;
-use sov_stf_runner::{from_toml_path, get_ledger_rpc, RollupConfig, StateTransitionRunner};
+use sov_stf_runner::{from_toml_path, RollupConfig, StateTransitionRunner};
 use tracing::{debug, Level};
 
 #[cfg(test)]
 mod test_rpc;
-
-#[cfg(feature = "experimental")]
-const TX_SIGNER_PRIV_KEY_PATH: &str = "../test-data/keys/tx_signer_private_key.json";
-
-// The rollup stores its data in the namespace b"sov-test" on Celestia
-// You can change this constant to point your rollup at a different namespace
-const ROLLUP_NAMESPACE: NamespaceId = NamespaceId(ROLLUP_NAMESPACE_RAW);
-
-/// Initializes a [`LedgerDB`] using the provided `path`.
-pub fn initialize_ledger(path: impl AsRef<std::path::Path>) -> LedgerDB {
-    LedgerDB::with_path(path).expect("Ledger DB failed to open")
-}
-
-// TODO: Remove this when sov-cli is in its own crate.
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-struct HexKey {
-    hex_priv_key: String,
-    address: String,
-}
-
-/// Configure our rollup with a centralized sequencer using the SEQUENCER_DA_ADDRESS
-/// address constant. Since the centralize sequencer's address is consensus critical,
-/// it has to be hardcoded as a constant, rather than read from the config at runtime.
-///
-/// If you want to customize the rollup to accept transactions from your own celestia
-/// address, simply change the value of the SEQUENCER_DA_ADDRESS to your own address.
-/// For example:
-/// ```rust,no_run
-/// const SEQUENCER_DA_ADDRESS: [u8;47] = *b"celestia1qp09ysygcx6npted5yc0au6k9lner05yvs9208"
-/// ```
-pub fn get_genesis_config() -> GenesisConfig<DefaultContext> {
-    let hex_key: HexKey = serde_json::from_slice(include_bytes!(
-        "../../test-data/keys/token_deployer_private_key.json"
-    ))
-    .expect("Broken key data file");
-    let sequencer_private_key = DefaultPrivateKey::from_hex(&hex_key.hex_priv_key).unwrap();
-    assert_eq!(
-        sequencer_private_key.default_address().to_string(),
-        hex_key.address,
-        "Inconsistent key data",
-    );
-    let sequencer_da_address = CelestiaAddress::from_str(SEQUENCER_DA_ADDRESS).unwrap();
-    create_demo_genesis_config(
-        100000000,
-        sequencer_private_key.default_address(),
-        sequencer_da_address.as_ref().to_vec(),
-        &sequencer_private_key,
-        &sequencer_private_key,
-    )
-}
 
 /// Main demo runner. Initialize a DA chain, and starts a demo-rollup using the config provided
 /// (or a default config if not provided). Then start checking the blocks sent to the DA layer in
@@ -158,53 +100,4 @@ async fn main() -> Result<(), anyhow::Error> {
     runner.run().await?;
 
     Ok(())
-}
-
-fn register_sequencer<DA>(
-    da_service: DA,
-    demo_runner: &mut App<
-        Risc0Verifier,
-        <DA::Spec as DaSpec>::ValidityCondition,
-        jupiter::BlobWithSender,
-    >,
-    methods: &mut jsonrpsee::RpcModule<()>,
-) -> Result<(), anyhow::Error>
-where
-    DA: DaService<Error = anyhow::Error> + Send + Sync + 'static,
-{
-    let batch_builder = demo_runner.batch_builder.take().unwrap();
-    let sequencer_rpc = get_sequencer_rpc(batch_builder, Arc::new(da_service));
-    methods
-        .merge(sequencer_rpc)
-        .context("Failed to merge Txs RPC modules")
-}
-
-fn register_ledger(
-    ledger_db: LedgerDB,
-    methods: &mut jsonrpsee::RpcModule<()>,
-) -> Result<(), anyhow::Error> {
-    let ledger_rpc = get_ledger_rpc::<SequencerOutcome<CelestiaAddress>, TxEffect>(ledger_db);
-    methods
-        .merge(ledger_rpc)
-        .context("Failed to merge ledger RPC modules")
-}
-
-#[cfg(feature = "experimental")]
-fn register_ethereum(
-    da_config: DaServiceConfig,
-    methods: &mut jsonrpsee::RpcModule<()>,
-) -> Result<(), anyhow::Error> {
-    use std::fs;
-
-    let data = fs::read_to_string(TX_SIGNER_PRIV_KEY_PATH).context("Unable to read file")?;
-
-    let hex_key: HexKey =
-        serde_json::from_str(&data).context("JSON does not have correct format.")?;
-
-    let tx_signer_private_key = DefaultPrivateKey::from_hex(&hex_key.hex_priv_key).unwrap();
-
-    let ethereum_rpc = get_ethereum_rpc(da_config, tx_signer_private_key);
-    methods
-        .merge(ethereum_rpc)
-        .context("Failed to merge Ethereum RPC modules")
 }
